Tests: Relocated some tests from main to common
