+++ arc/mainTest.py
@@ -9,17 +9,16 @@ from __future__ import (absolute_import, division, print_function, unicode_liter
 import unittest
 import os
 import shutil
-import time
 
 from rmgpy import settings
 from rmgpy.data.rmg import RMGDatabase
 from rmgpy.species import Species
 from rmgpy.molecule.molecule import Molecule
 
-from arc.main import ARC, read_file, get_git_commit, time_lapse, check_ess_settings
+from arc.main import ARC
 from arc.species.species import ARCSpecies
 from arc.settings import arc_path, servers
-from arc.arc_exceptions import InputError, SettingsError
+from arc.arc_exceptions import InputError
 
 ################################################################################
 
@@ -274,63 +273,6 @@ class TestARC(unittest.TestCase):
         # delete the generated library from RMG-database
         os.remove(new_thermo_library_path)
 
-    def test_check_ess_settings(self):
-        """Test the check_ess_settings function"""
-        ess_settings1 = {'gaussian': [self.servers[0]], 'molpro': [self.servers[1], self.servers[0]],
-                         'qchem': [self.servers[0]]}
-        ess_settings2 = {'gaussian': self.servers[0], 'molpro': self.servers[1], 'qchem': self.servers[0]}
-        ess_settings3 = {'gaussian': self.servers[0], 'molpro': [self.servers[1], self.servers[0]],
-                         'qchem': self.servers[0]}
-        ess_settings4 = {'gaussian': self.servers[0], 'molpro': self.servers[1], 'qchem': self.servers[0]}
-        ess_settings5 = {'gaussian': 'local', 'molpro': self.servers[1], 'qchem': self.servers[0]}
-
-        ess_settings1 = check_ess_settings(ess_settings1)
-        ess_settings2 = check_ess_settings(ess_settings2)
-        ess_settings3 = check_ess_settings(ess_settings3)
-        ess_settings4 = check_ess_settings(ess_settings4)
-        ess_settings5 = check_ess_settings(ess_settings5)
-
-        ess_list = [ess_settings1, ess_settings2, ess_settings3, ess_settings4, ess_settings5]
-
-        for ess in ess_list:
-            for soft, server_list in ess.items():
-                self.assertTrue(soft in ['gaussian', 'molpro', 'qchem'])
-                self.assertIsInstance(server_list, list)
-
-        with self.assertRaises(SettingsError):
-            ess_settings6 = {'nosoft': ['server1']}
-            check_ess_settings(ess_settings6)
-        with self.assertRaises(SettingsError):
-            ess_settings7 = {'gaussian': ['noserver']}
-            check_ess_settings(ess_settings7)
-
-    def test_time_lapse(self):
-        """Test the time_lapse() function"""
-        t0 = time.time()
-        time.sleep(2)
-        lap = time_lapse(t0)
-        self.assertEqual(lap, '00:00:02')
-
-    def test_get_git_commit(self):
-        """Test the get_git_commit() function"""
-        git_commit = get_git_commit()
-        # output format: ['fafdb957049917ede565cebc58b29899f597fb5a', 'Fri Mar 29 11:09:50 2019 -0400']
-        self.assertEqual(len(git_commit[0]), 40)
-        self.assertEqual(len(git_commit[1].split()), 6)
-
-    def test_read_file(self):
-        """Test the read_file() function"""
-        restart_path = os.path.join(arc_path, 'arc', 'testing', 'restart(H,H2O2,N2H3,CH3CO2).yml')
-        input_dict = read_file(restart_path)
-        self.assertIsInstance(input_dict, dict)
-        self.assertTrue('reactions' in input_dict)
-        self.assertTrue('freq_level' in input_dict)
-        self.assertTrue('use_bac' in input_dict)
-        self.assertTrue('ts_guess_level' in input_dict)
-        self.assertTrue('running_jobs' in input_dict)
-
-        with self.assertRaises(InputError):
-            read_file('nopath')
 
     def test_determine_model_chemistry(self):
         """Test determining the model chemistry"""
